mtest: Split neighborhood collective tests
Split them into multiple files. We will be testing
various algorithms for each of these neighborhood
collectives and therefore, they are better placed
in separate files.
